+++ axelor-business-production/src/main/java/com/axelor/apps/businessproduction/service/ManufOrderServiceBusinessImpl.java
@@ -31,6 +31,7 @@ import com.axelor.apps.production.service.manuforder.ManufOrderServiceImpl;
 import com.axelor.apps.production.service.manuforder.ManufOrderWorkflowService;
 import com.axelor.apps.production.service.operationorder.OperationOrderService;
 import com.axelor.exception.AxelorException;
+import com.axelor.inject.Beans;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 import java.lang.invoke.MethodHandles;
@@ -99,6 +100,11 @@ public class ManufOrderServiceBusinessImpl extends ManufOrderServiceImpl {
         super.createManufOrder(
             product, qty, priority, isToInvoice, company, billOfMaterial, plannedStartDateT);
 
+    AppProductionService appProductionService = Beans.get(AppProductionService.class);
+    if (!appProductionService.isApp("production")
+        || !appProductionService.getAppProduction().getManageBusinessProduction()) {
+      return manufOrder;
+    }
     manufOrder.setIsToInvoice(isToInvoice);
 
     return manufOrder;