+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/batch/MailBatchServiceHR.java
@@ -22,6 +22,7 @@ import com.axelor.apps.base.db.MailBatch;
 import com.axelor.apps.base.db.repo.MailBatchRepository;
 import com.axelor.apps.base.exceptions.IExceptionMessage;
 import com.axelor.apps.base.service.batch.MailBatchService;
+import com.axelor.apps.hr.service.app.AppHumanResourceService;
 import com.axelor.db.Model;
 import com.axelor.exception.AxelorException;
 import com.axelor.exception.db.repo.TraceBackRepository;
@@ -32,6 +33,10 @@ public class MailBatchServiceHR extends MailBatchService {
   @Override
   public Batch run(Model batchModel) throws AxelorException {
     Batch batch = super.run(batchModel);
+    if (!Beans.get(AppHumanResourceService.class).isApp("employee")) {
+      return batch;
+    }
+
     MailBatch mailBatch = (MailBatch) batchModel;
 
     switch (mailBatch.getActionSelect()) {