+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/service/ProductServicePurchaseImpl.java
@@ -24,6 +24,7 @@ import com.axelor.apps.base.service.ProductServiceImpl;
 import com.axelor.apps.base.service.ProductVariantService;
 import com.axelor.apps.base.service.administration.SequenceService;
 import com.axelor.apps.base.service.app.AppBaseService;
+import com.axelor.inject.Beans;
 import com.google.inject.Inject;
 
 public class ProductServicePurchaseImpl extends ProductServiceImpl {
@@ -41,6 +42,9 @@ public class ProductServicePurchaseImpl extends ProductServiceImpl {
   @Override
   public void copyProduct(Product product, Product copy) {
     super.copyProduct(product, copy);
-    copy.setSupplierCatalogList(null);
+
+    if (Beans.get(AppBaseService.class).isApp("purchase-request")) {
+      copy.setSupplierCatalogList(null);
+    }
   }
 }