+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/PartnerSaleService.java
@@ -19,7 +19,9 @@ package com.axelor.apps.sale.service;
 
 import com.axelor.apps.base.db.Partner;
 import com.axelor.apps.base.service.PartnerService;
+import com.axelor.apps.sale.service.app.AppSaleService;
 import com.axelor.db.JPA;
+import com.axelor.inject.Beans;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
@@ -28,6 +30,12 @@ public class PartnerSaleService extends PartnerService {
 
   @Override
   public List<Long> findPartnerMails(Partner partner) {
+
+    if (!Beans.get(AppSaleService.class).isApp("sale")) {
+
+      return super.findPartnerMails(partner);
+    }
+
     List<Long> idList = new ArrayList<Long>();
 
     idList.addAll(this.findMailsFromPartner(partner));
@@ -45,6 +53,11 @@ public class PartnerSaleService extends PartnerService {
 
   @Override
   public List<Long> findContactMails(Partner partner) {
+
+    if (!Beans.get(AppSaleService.class).isApp("sale")) {
+      return super.findContactMails(partner);
+    }
+
     List<Long> idList = new ArrayList<Long>();
 
     idList.addAll(this.findMailsFromPartner(partner));