+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/StockRulesServiceSupplychainImpl.java
@@ -38,6 +38,7 @@ import com.axelor.apps.stock.db.StockLocationLine;
 import com.axelor.apps.stock.db.StockRules;
 import com.axelor.apps.stock.db.repo.StockRulesRepository;
 import com.axelor.apps.stock.service.StockRulesServiceImpl;
+import com.axelor.apps.supplychain.service.app.AppSupplychainService;
 import com.axelor.auth.AuthUtils;
 import com.axelor.exception.AxelorException;
 import com.axelor.exception.db.repo.TraceBackRepository;
@@ -79,6 +80,11 @@ public class StockRulesServiceSupplychainImpl extends StockRulesServiceImpl {
   public void generatePurchaseOrder(
       Product product, BigDecimal qty, StockLocationLine stockLocationLine, int type)
       throws AxelorException {
+
+    if (!Beans.get(AppSupplychainService.class).isApp("supplychain")) {
+      super.generatePurchaseOrder(product, qty, stockLocationLine, type);
+      return;
+    }
 
     StockLocation stockLocation = stockLocationLine.getStockLocation();