+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Improvements
 - PurchaseRequestLine : cacheable removed for this entity
 - BANKSTATEMENT : on copy reset statusSelect
+- Check and modify all service code according to app configuration
+
 ## Bug Fixes
 
 ## [5.1.14] - 2020-01-17